# api-docs_1.2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Discodeit API 문서",
    "description": "Discodeit 프로젝트의 Swagger API 문서입니다.",
    "version": "1.2"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "로컬 서버"
    }
  ],
...
  "paths": {
    "/api/users": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "전체 User 목록 조회",
        "operationId": "findAll",
        "responses": {
          "200": {
            "description": "User 목록 조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "User"
        ],
        "summary": "User 등록",
        "operationId": "create",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "userCreateRequest": {
                    "$ref": "#/components/schemas/UserCreateRequest"
                  },
                  "profile": {
                    "type": "string",
                    "format": "binary",
                    "description": "User 프로필 이미지"
                  }
                },
                "required": [
                  "userCreateRequest"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User가 성공적으로 생성됨",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          },
          "400": {
            "description": "같은 email 또는 username를 사용하는 User가 이미 존재함",
            "content": {
              "*/*": {
                "example": "User with email {email} already exists"
              }
            }
          }
        }
      }
    },
    "/api/readStatuses": {
      "get": {
        "tags": [
          "ReadStatus"
        ],
        "summary": "User의 Message 읽음 상태 목록 조회",
        "operationId": "findAllByUserId",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "description": "조회할 User ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message 읽음 상태 목록 조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReadStatusDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ReadStatus"
        ],
        "summary": "Message 읽음 상태 생성",
        "operationId": "create_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadStatusCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "이미 읽음 상태가 존재함",
            "content": {
              "*/*": {
                "example": "ReadStatus with userId {userId} and channelId {channelId} already exists"
              }
            }
          },
          "201": {
            "description": "Message 읽음 상태가 성공적으로 생성됨",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReadStatusDTO"
                }
              }
            }
          },
          "404": {
            "description": "Channel 또는 User를 찾을 수 없음",
            "content": {
              "*/*": {
                "example": "Channel | User with id {channelId | userId} not found"
              }
            }
          }
        }
      }
    },
...
    "/api/users/{userId}": {
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "User 삭제",
        "operationId": "delete",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "삭제할 User ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "User를 찾을 수 없음",
            "content": {
              "*/*": {
                "example": "User with id {id} not found"
              }
            }
          },
          "204": {
            "description": "User가 성공적으로 삭제됨"
          }
        }
      },
      "patch": {
        "tags": [
          "User"
        ],
        "summary": "User 정보 수정",
        "operationId": "update",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "수정할 User ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "userUpdateRequest": {
                    "$ref": "#/components/schemas/UserUpdateRequest"
                  },
                  "profile": {
                    "type": "string",
                    "format": "binary",
                    "description": "수정할 User 프로필 이미지"
                  }
                },
                "required": [
                  "userUpdateRequest"
                ]
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "같은 email 또는 username를 사용하는 User가 이미 존재함",
            "content": {
              "*/*": {
                "example": "user with email {newEmail} already exists"
              }
            }
          },
          "200": {
            "description": "User 정보가 성공적으로 수정됨",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          },
          "404": {
            "description": "User를 찾을 수 없음",
            "content": {
              "*/*": {
                "example": "User with id {userId} not found"
              }
            }
          }
        }
      }
    },
...
    "/api/binaryContents": {
      "get": {
        "tags": [
          "BinaryContent"
        ],
        "summary": "여러 첨부 파일 조회",
        "operationId": "findAllByIdIn",
        "parameters": [
          {
            "name": "binaryContentIds",
            "in": "query",
            "description": "조회할 첨부 파일 ID 목록",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "format": "uuid"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "첨부 파일 목록 조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BinaryContentDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/binaryContents/{binaryContentId}/download": {
      "get": {
        "tags": [
          "BinaryContent"
        ],
        "summary": "파일 다운로드",
        "operationId": "download",
        "parameters": [
          {
            "name": "binaryContentId",
            "in": "path",
            "description": "다운로드할 파일 ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "파일 다운로드 성공",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "UserCreateRequest": {
        "type": "object",
        "description": "User 생성 정보",
        "properties": {
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "BinaryContentDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fileName": {
            "type": "string"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "contentType": {
            "type": "string"
          }
        }
      },
      "UserDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "profile": {
            "$ref": "#/components/schemas/BinaryContentDTO"
          },
          "online": {
            "type": "boolean"
          }
        }
      },
      "ReadStatusCreateRequest": {
        "type": "object",
        "description": "Message 읽음 상태 생성 정보",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "channelId": {
            "type": "string",
            "format": "uuid"
          },
          "lastReadAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ReadStatusDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "channelId": {
            "type": "string",
            "format": "uuid"
          },
          "lastReadAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "UserUpdateRequest": {
        "type": "object",
        "description": "수정할 User 정보",
        "properties": {
          "newUsername": {
            "type": "string"
          },
          "newEmail": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          }
        }
      },
...
    }
  }
}